requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
pillow==9.4.0
requests==2.28.2
openai==0.27.2
pydub==0.25.1
python-magic==0.4.27
gunicorn==20.1.0
google-cloud-vision==3.4.0
google-cloud-speech==2.21.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
